+++ evalml/pipelines/regression/__init__.py
@@ -3,4 +3,5 @@ from .linear_regression import LinearRegressionPipeline
 from .random_forest import RFRegressionPipeline
 from .catboost import CatBoostRegressionPipeline
 from .xgboost_regression import XGBoostRegressionPipeline
+from .extra_trees import ETRegressionPipeline
 from .baseline_regression import BaselineRegressionPipeline, MeanBaselineRegressionPipeline